Added StreamData class for reading / writing samples. Basic calc.
Basic calc is only to support simple math like 1e6 or 325e6/100 in input
file.

Now I'm working on massaging StreamData samples into something that can
be used with tx_streamer / rx_streamer. Will likely be switching back to
std::vector in spite of the fact that maps are so much nicer to work
with.