Segfault is mitigated.
Not sure what caused it in the end. I just backtracked, and reverted to
the existing txrx file io example and re-performed all of my changes
incrementally, and it ended up working.

I believe shared pointers were playing a role, because I could inspect
the memory of the object where the segfault was occuring, but inside its
own methods, it was unable to dereference 'this' which was really odd.