Added UHD support for common-mode settings.
Made adjustments for multi-crimson in:

set_rx_freq()
set_rx_gain()
set_tx_freq()
set_tx_gain()

We are unable to use common-mode for rx/freq/lna because those settings
are currently not timed.

NB: With this change, UHD only checks that the mask is already set prior
to writing individual or common-mode settings. If UHD is to set the
common-mode setting itself, there will need to be some formal criteria
for it to initialize / deinitialize common mode. My thoughts are, that
that should be up to the tx_streamer / rx_streamer classes, in the
constructors (or initializers) / and destructors (deinitializers).