Verified Vita header at beginning of UDP Frame with Wireshark.
For some reason wireshark's vrt49 dissector does not seem to parse the packet
properly, possibly because it's a large UDP frame. However, I was able to
manually verify that all VITA header bytes were sent in the raw UDP payload.

Originally, my code used Ettus' UHD vrt stuff, and originally I thought that
was somehow broken, which may be the case, but it appears to have just been
conflated with the wireshark issue.

I see no point in reverting the changes in this commit, although it is
slightly more ugly than before.