Add hysteresis for convergence. Check for null _crimson_tng_impl
The root cause of the crash before, while waiting for convergence in init, was
that the _crimson_tng_impl pointer variable was being dereferenced before it
was set.

I explicitly initialized it to NULL and perform a NULL check to avoid this.

This likely occurred as a result of enabling the SoB convergence full-time,
because, before, it would only be enabled when a burst was SoB was transmitted.

Also, added hysteresis for the convergence levels, so that it must be within
an even lower threshold to be converged, and must be outside of a higher
threshold to be diverged, a-la Schmitt Trigger.