Fix SMA calculation. Give differentiator initial values. Put SoB code into separate methods.
As it turns out, the Simple Moving Average (SMA) calculation that was on the wikipedia was originally wrong. There was some changes on wikipedia that
went in April 17-19 or so. I updated my SMA code to reflect those changes. Added unit tests for SMA.

"I knew exactly what was wrong when I heard sys13 and 0x7"

... PPS Test code was conflating Start of Burst (SoB) / Time Diff calculations. What happened, was that UHD would send Crimson a monotonic time. Then shortly after, that time would be set to zero. Subsequent Time Diff packets would produce PID Process Variable (PV) values that were WAY off, on the order of hundreds of thousands of seconds. Subsequently, rather than having a PID Control Variable of a fraction of a second (representing network latency and process latency), the PID control Variable would be massive.

Specifically for Start of Burst, send would go into a sleep for CV - 0.001 seconds, which would stop TX. So the problem wasn't that settings were not set in #3836, it was that there was no data being sent.

This task is intended to remove PPS Test routines from the server code.