Only perform PID calculations for 0th txstreamer instance.
Only 1 PID should exist per physical Crimson unit.

Part of that task should be to move the PID loop from the txstreamer class to
the device class. However, the primary obstacle with this is, since the time
diff mechanism is so tightly coupled with flow control, doing so seems
impractical, currently.

A second obstacle with the second item is that (in the future), we can predict
that we will be using multiple crimson units in parallel for phased arrays and
things of that nature, so the decision as to whether or not a physical Crimson
unit has a PID controller associated with it will also require referencing
some kind of map based on IP address or MAC address.

A third obstacle with the second item is that if there are multiple OS
processes interacting with the same physical Crimson unit, each of those will
have its own copy of the libuhd.so shared library, each of which has it's own,
conflicting, PID controller. The solution in that case, is to either

a) use a type of 'cookie' that is stored in some location in shared memory, or
b) use a type of 'cookie' that is stored on Crimson

Above, the first solution addresses the case that multiple processes *on the
same host compuer* are accessing a single Crimson unit. The second solution
addresses the case that multiple processes on *the same or potentially
different host computers* are accessing a single Crimson unit. Obstacle number
3 is to be considered a very advanced use case, and gets into the realm of
distributed shared resources, locking, and so on.

In either case, obstacle number 3 is also something that we must push back to
fix at a later date, potentially after reworking the way that we
conceptualize Crimson as a shared resource on a distributed system.