Workaround for bug in g++
This was due to a compiler bug in G++.

I saw the following output when using my txrx_to_from_file_loopback example.

<pre>
pure virtual method called
terminate called without an active exception
</pre>

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=52972

<pre>
The standard says "Member functions can be called from a constructor (or destructor) of an abstract class; the effect of making a virtual call to a pure virtual function directly or indirectly for the object being created (or destroyed) from such a constructor (or destructor) is undefined."
</pre>

Since this function was not being called from the context of any constructor, and since I have cleanly stepped through the _implementation_ of this (very non-pure-virtual) function, this is an obvious g++ compiler bug.

I'm submitting a workaround for it and will refactor recv() in a separate ticket.