Buffer contents when nsamps_per_channel lt vita len
Previously, if recv() was called with nsamps_per_buff < vita packet_size,
we would return only nsamps_per_buff. Subsequent reads would start in the
middle of a vita packet, and we would mistakenly parse a non-existent vita
header.

This change ensures that we buffer remaining vita packet data when
nsamps_per_buff < vita packet_size and that we return that data and the
corresponding metadata upon subsequent reads.