flow_control class passing all unit tests at required sample rate.
Like with a number of unit tests, I've used a kind of mathematical
induction.

Test #1: Test empty buffer

After initialization, the flow controller should report that there are
zero samples in the buffer and that the sample rate is equal to the
nominal sample rate.

Test #2: Test time until first send

The first time send is called, with an empty buffer, there should be
zero delay.

Test #3: Test buffer level after first update

Immediately after the first update, there should be exactly the same
number of items in the buffer as sent to it. (this was actually off by
1, but the number was 2225). Likely a rounding issue.

Test #4: Test time until second send

The time to wait before sending the first update is zero, but the time
to wait before sending the second update should be slightly greater than
that. Here, we are trying to fill the buffer quickly by increasing the
sample rate. With a high sample rate, the delay is very short.

Time to wait before sending is inverseley proportional to sample rate.

Test #5: Test inter-pid-sample convergence

This test focused on the transient condition of starting at zero and
then ramping up the buffer level (via increased sample rate). The trick
here was only updating the Control Variable of the PID controller once
per PID sample period. Otherwise, the sample rate was changed according
to the formula below:

sample_rate = nominal_sample_rate + ( cv - pv ) * 2 / pid_sample_rate;

There are likely a number of kinks to work out, and I would like to
graph the buffer level over time as well.