Fix Ku parameter to pidc constructor. Add documentation.
Previously I had been doing a double-conversion for calculating Kp.

Since the error signal was already in samples, I only needed to use a
simple floating point number for Ku (which affects Kp, Ki, and Kd), but
for some reason I had converted it to number of samples twice, which
would have squared the error signal.

That would explain why it was saturating.

In any case, that is fixed.

Also commented code.

Current test results are extremely promising and show fast convergence
to 80 % @ 54 MSps, within 0.01s with near perfect stability. The only
variance is due to MTU size.