In-band request and response for time diff working
Note: somewhere (potentially in UHD or the kernel), udp packets are filtered out if their source port in the response packet
is not the same as the dest port in the request packet. We verified that menaually by setting the port in the state tree. Will be built-in to next
FPGA image.

I wanted to also add tree properties for qa/{fifo_lvl,uflow,oflow} but for some reason it does not seem to be working, even though it is verified
on the server. For now, I just commented them out in UHD.