Reverse order of fifo levels
The time diff response packet is defined as shown below

struct time_diff_resp {
	int64_t tv_sec;
	int64_t tv_tick;
	uint16_t fifo[ CRIMSON_TNG_TX_CHANNELS ];
	struct time_diff_uoflow uoflow[ CRIMSON_TNG_TX_CHANNELS ];
};

The implicit assumption was that channel fifo levels were sent in-order ( e.g. A, B, C, D ), but they were actually sent in reverse order because
the fifo levels are treated as single, 64-bit word in Verilog, and our Verilog code is implicitly big-endian.

It's a bit counter-intuitive, because a C compiler will automatically order the bytes for you when you say

uint64_t X = D << 48 | C << 32 | B << 16 | A << 0;

That is, on a big-endian machine, D would be in the lowest addressable byte of X and A would be in the highest addressable byte of X.
That way, if you do htonll( x ), nothing needs to be done.

On a little-endian machine, D would be in the highest addressable byte of X and A would be in the lowest addressable byte of X.
That way, if you do htonll( x ), it is converted into network (big-endian) byte order.

In Verilog, there is no 'htonll', i.e. there is no built-in serialization for multi-byte types. So if a uint64_t was created as if it were little
endian, then it will be sent the opposite way.

This is a software work-around for that.