Workaround for GNU Radio first call with zero samples.
GNU Radio calls tx_streamer::send() first with nsamps_per_buff set to 0. I do not know why this is the case, but we do not send 0-length packets
and we return quickly from send when nsamps_per_buff is 0.

This change is a temporary workaround until we validate our FPGA to work with 0-sample VRT49 packets.