Dont set time on startup, dont use abs(pv) for err
UHD now does not set crimson time but instead uses the time difference
on start as a constant offset. Crimson set time should only be used
when a GPS reference is present with pps

Also in this update there is a bug fix where the convergence code uses
the absolute value of PV for error. This in effect doubles the error