choose_dsp_nco_freq fix for 260 MSps plus unit tests
There should most definitely be more unit tests put into this, but for the time being, I think this solves one of our customer's issues.

The problem was that the full supported sample rate / bandwidth was not included as a catch-all, and only the upper-half of the bandwidths. When not considering the catch-all case of 325 MHz of
contiguous bandwidth in a single channel, it made sense to only consider the RHS frequency ranges since they were (by definition) non-contiguous with the LHS. However, it is fundamental to note,
that in a direct conversion system, the entire sample rate is available for bandwidth because we are considering complex sample data (i.e. we are not limited to providing a real signal only at
baseband).