Use more of upstream UHD code for RX Streamer
This was fairly straightforward to achieve and is the first step to achieving
compatibility with UHD upstream, if we would like to rebase off of the latest
code, that should be at least a little easier now.

There were a couple of gotchas.

1)  I was unable to use the existing udp_zero_copy class because it fixes the
remote IP address and port (on the device side) rather than the local IP
address and port (on the host side). The reason for this is a bit funny. With
the USRP2, the device learns the IP address and port of the host by extracting
it out of a dummy packet that is sent to the device's RX port on init. With
Crimson, we program the host's RX IP and port via the web or by setting
properties.

2) I had to append "_item32_le" instead of "_item32_be" in the wire format.

It would be really great to do something similar with TX, with the addition
that we would still need to have threads for buffer management & clock
synchronization.