Add double-sided band for higher sample rates
Our product is supposed to be able to deliver 260 MSps on channel A+B and half of that on 130 MSps on channel C+D.

Prior to this change, we were only using 1/2 the available bandwidth on each channel because we had only specified the upper sideband of the spectrum. Since we are working with direct-conversion (and therefore the lower sideband does not necessarily need to mirror the upper sideband), our effective information bandwidth doubles.