First commit for switching to super_send_packet_handler
This is nearly finished.
One tricky part is that we currently bundle our uoflow and buffer_level statistics are bundled with the time diff.
That has the advantage that more information is contained in one packet and, therefore, there is less overhead with collective updates.
However, the disadvantage is that the buffer level information, which is really a concern of the streamer, is not separable from the time diff information, which is a concern of the usrp object.
We only did it this way to quickly get a working system out the door, but as a result, it forces me to keep the listener model, which I just did away with prematurely.