We will have to drop the VRT49 TSI field on TX I think.
There is no was for us to change the tsi_type / has_tsi fields dynamically (at least re-using upstream code), so I say we remove it.

Use the reserved bit in the vita header to determine whether we have a TX SoB and just use the free-running counter.

Make it so the free-running counter (for TX) does not wrap at 1s, but have wrap it at 64-bits.