This commit combines a few changes
There were two changes that were previously done but lost in rebasing to ettus/master

* disable rx stream before enabling it
* power on and enable rx stream before sending out any stream command that is not STOP_CONTINUOUS

The first one has no real side effects and just signals to the FPGA that it should empty the RX buffer.

The second one is part of a fairly nasty set of workarounds for both TX and RX that have to do with broken reference counting in shared pointers. I believe this is a UHD-wide bug and not specific to crimson_tng, but I'm not certain. Other USRP targets perform similar workarounds - e.g. sending "mini-eob" to signal the streamer to tear itself down.

The last change just allows some of the set_tx and set_rx methods in multi_crimson_tng to recognize ALL_CHANS. Note, that we are breaking spec with multi_usrp which specifies that the default channel to take effect with is channel zero. However, doing so seems to be mostly desireable behaviour.