Combine multi_crimson_tng.cpp into multi_usrp.cpp
In the future, we should switch to 100% reuse of multi_usrp.cpp. For the time being, we have overwritten the upstream ettus methods for setting & getting frequency and gain with out own.
However, most of that business logic should go on the server eventually. We should be able to copy the code as is, more or less, if we use c++ for the server.

Some notable achievements:

* fixed a circular reference in crimson_tng_tx_streamer which allowed us to hit that destructor properly.
* added a few more properties in crimson_tng_impl.cpp that let us become closer to parity with upstream ettus
* in the future, we should still break out gain elements into individual properties as well as frequency elements (lo's & nco offsets).