dynamically stop and start the pillage thread
We have to do something on end-of-burst, because that's the only way that some applications inform us that they are closing.

E.g. benchmark_thread would only send end of burst to indicate that they were finished. Without "interrupting" threads embedded within the send_streamer, benchmark_rate would just hang waiting for threads to
join.

I have a feeling that another thread is hanging - the "bm thread". This used to be used for buffer management (i.e. flow control) but now it's exclusively used for keeping track of the time diff. We *really*
should move the code that keeps track of the time diff into io_impl.cpp, and I did have another branch to do that but just haven't gotten to it yet. I believe the "bm thread" is hanging when we would like to
close gnu radio, and until that is resolved we need to keep the atexit hack.