Use pre-packed SC16 and bypass converting samples (scatter/gather)
We can have the data in one buffer and the VRT header in the other. If
the data is already packed in a sc16 format (32-bit storage, 16 for I
and 16 for Q), then we can bypass the costly convert function and
directly pass those buffers to sendmmsg which supports scatter/gather
buffers.