cyan_64t: device stable so add to master
This change pulls cyan_64t changes into master, as it is now stable.

Previously, the master branch was being used as a product branch for crimson_tng.
master-testing was being used as a product branch for cyan_64t.

Since we do not intend to use product branches and would like master to support
all of our products, we would like to convert this to a feature branch and
then merge with master.

Signed-off-by: Christopher Friedt <chris.f@pervices.com>