cyan_16t: platform is stable
Previously, the master branch was being used as a product branch for crimson_tng.
ops/vvv/master-testing was being used as a product branch for cyan_16t.

Since we do not intend to use product branches and would like master to support
all of our products, we would like to convert this to a feature branch and
then merge with master.

The same will be done for master-testing, which was being used as a
product branch for cyan_64t. The master-testing branch should nominally be used
as staging (i.e. for 'soak' testing), and that will be corrected in a
subsequent PR, as cyan_64t is now stable.

Signed-off-by: Christopher Friedt <chris.f@pervices.com>