CYAN_8R: temporary fix for initialization failing
When initializing the device, uhd will try to use tx hooks that
shouldn't exist for 8r. The temporary fix is to have the number of
transmit and receive channel match. The long term solution is to make
the part of the code that tries to access the tx stuff not run number of
tx channels = 0