SEND: querry buffer level while waiting for cleanup
Previously the streamer would wait until it thought the buffer was
empty based on the same buffer level prediction done by the main
program. Now it will check the buffer until it is empty. This is worse
performance wise, but is done to avoid potential issues from finishing
cleanup with samples in the buffer. Also performance does not matter at
this stage