CYAN: remove set/get every property when initializing
It serves no purpose since the properties are either WO or will override
get to retrieve the value from the server anyway so the value obtained
by this get is never used.

Previously it would get the current value of the property on the server,
the set the value in UHD's copy of the state tree to that value. However
this does nothing since the value in UHD's state tree is never accessed
instead the value is retrieved from the server