TEST_TX_TRIGGER: throw error if samples are not sent + correct num triggers
Throw error if timeout when sending samples: triggers are to far apart
or triggers are not being activated, something is very wrong, test
should fail

Previously the number of triggers argument was actually just the number
of times the buffer level should be polled. Now that argument is
actually the number of triggers to complete. NOTE: this will result in
the automated tests taking much longer with the same value for said
argument