CLOCK SYNC: moving info to shared object
Move clock sync info that needs to be communicated between the
device managed clcok sync thread and the streamer.

This eliminates binding the device get_time_now function to the send
streamer which might be a source of isntability on exit if the device
gets destructed first. It also moves some duplicate code between Cyan
and Crimson to a common file