IO_URING fixing not waking from sleep
If you submit an io_uring request from one thread (such as when creating
an rx streamer), and process it in another (such as when calling recv)
the originall thread won't wake from sleep.

Ideally I'd figure out how to prevent issues with sleep in the
submitting threads but I have been unable to succeed in that.