RECV: add instructions for giving non root access to preempt check
Previously it would tell the user to run it as root so the check could
run, now it tells the user how to let non root perform the check.

I also changed it so that if the file does not exist it will the the
user instructions on how to mount debugfs. Previously it would silently
be unable to check if preemption is disabled.