TX: Changing message queue to lockless
Locking in the queue used to send messages to inform the user of
overflows and underflows could cause programs to hang sometimes. The new
system is lockless.

It does have the issue of reading from the messages queue from multiple
threads could cause duplicate messages to be read, but I have deemed
that worth it since the impact of that is minor, and users don't have a
reason to read from the queue from more than one thread. If it becomes a
problem then a lock can be added to the read side only.